Reorganize src structure.
